player: allow for C plugins
This basically reuses the scripting infrastructure. Because of that, it
uses --lua and ~/.mpv/lua/ to load C plugins. If a "script" has the file
extension ".so", it will be opened with dlopen(), and the following
function is called:

    int mpv_dlopen(mpv_handle *handle);

mpv_handle is as in the client API uses it (libmpv/client.h). The
function runs in its own thread, and the plugin is considered as
"loaded" as long as this function doesn't return. If a value other than
0 is returned, an error is signaled to the user.

Note that this needs to be explicitly enabled at compilation. For one,
enabling export for certain symbols from an executable seems to be quite
toolchain-specific. It might not work outside of Linux.